app/requirements.txt
flask==3.0.0
boto3==1.34.0
psycopg2-binary==2.9.9
geopandas==0.14.4
shapely==2.0.6
python-dotenv==1.0.0
